+++ backend/src/main/java/com/tracker/backend/models/DailyGoal.java
@@ -12,10 +12,10 @@ public class DailyGoal {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer goalId;
+    private Long goalId;
 
     @Column(name = "user_id", nullable = false)
-    private Integer userId;
+    private Long userId;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
@@ -40,7 +40,7 @@ public class DailyGoal {
         // Default constructor
     }
 
-    public DailyGoal(Integer userId, Category category, Integer goalValue, DayOfWeek dayOfWeek, String goalTime) {
+    public DailyGoal(Long userId, Category category, Integer goalValue, DayOfWeek dayOfWeek, String goalTime) {
         this.userId = userId;
         this.category = category;
         this.goalValue = goalValue;
@@ -50,19 +50,19 @@ public class DailyGoal {
     }
 
     // Getters and Setters
-    public Integer getGoalId() {
+    public Long getGoalId() {
         return goalId;
     }
 
-    public void setGoalId(Integer goalId) {
+    public void setGoalId(Long goalId) {
         this.goalId = goalId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
